requirements.txt
aiogram==2.25.1
python-dotenv==0.20.0
gino==1.0.1
SQLAlchemy==1.3.24
asyncpg==0.26.0
aioschedule==0.5.2
pandas~=2.0.1
DateTime~=5.1
numpy~=1.23.5
opencv-python~=4.7.0.72
matplotlib~=3.7.1
scikit-learn~=1.2.2
keras~=2.12.0
tensorflow~=2.12.0